fix tests and fix struct::agg_list
