+++ polars/polars-lazy/Cargo.toml
@@ -50,7 +50,7 @@ cum_agg = ["polars-core/cum_agg"]
 interpolate = ["polars-core/interpolate"]
 rolling_window = ["polars-core/rolling_window"]
 rank = ["polars-core/rank"]
-diff = ["polars-core/diff"]
+diff = ["polars-core/diff", "polars-ops/diff"]
 pct_change = ["polars-core/pct_change"]
 moment = ["polars-core/moment"]
 list = ["polars-ops/list"]
