+++ polars/polars-ops/Cargo.toml
@@ -23,6 +23,6 @@ dtype-struct = ["polars-core/dtype-struct"]
 dtype-u8 = ["polars-core/dtype-u8"]
 object = ["polars-core/object"]
 to_dummies = []
-list_to_struct = ["polars-core/dtype-struct"]
+list_to_struct = ["polars-core/dtype-struct", "list"]
 list = []
 diff = []
